requirements.txt
fastapi
uvicorn[standard]
reportlab
PyPDF2
python-dotenv